Enabled Window Reizable
Now the window is resizable. The canvas size is still fixed to 1920x1080, though this can be changed in the code.
In the future maybe I'll set the canvas size to match the window size (a bit tricky to not crash in some fast testing),
or some other thing to allow moving the canvas or something.